backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]
